requirements.txt
# Web framework (TensorFlow-compatible)
fastapi==0.95.2
uvicorn==0.22.0

# ML
tensorflow==2.13.0
numpy==1.24.4
pillow==10.2.0
scikit-learn==1.3.2

# Experiment tracking
mlflow==2.10.0
pytest==7.4.4
